TASK: Move `GlobalState` up to `@neos-project/neos-ui-redux-store` and merge with `State`
