+++ packages/neos-ui-redux-store/src/UI/EditPreviewMode/index.ts
@@ -1,6 +1,7 @@
 import {action as createAction, ActionType} from 'typesafe-actions';
 
-import {actionTypes as system, InitAction, GlobalState} from '../../System';
+import {actionTypes as system, InitAction} from '../../System';
+import {GlobalState} from '../..';
 
 export type State = string;
