+++ packages/neos-ui-redux-store/src/UI/RightSideBar/index.ts
@@ -1,7 +1,8 @@
 import produce from 'immer';
 import {action as createAction, ActionType} from 'typesafe-actions';
 
-import {actionTypes as system, InitAction, GlobalState} from '../../System';
+import {actionTypes as system, InitAction} from '../../System';
+import {GlobalState} from '../..';
 
 export interface State extends Readonly<{
     isHidden: boolean;
